Switch to fixed fabber ratios
Due to the Bugs having a singular land fabber, the assumptions of the AI system break down, meaning the NeedBasicBotFabber and NeedBasicVehicleFabber checks cannot be relied upon. This alternate approach removes one element for AI personalities to modify, but the fix is relatively critical for the AI to operate correctly.